+++ exporters/stdout/exporter.go
@@ -15,11 +15,13 @@
 package stdout // import "go.opentelemetry.io/otel/exporters/stdout"
 
 import (
+	"context"
+
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/sdk/export/metric"
 	exporttrace "go.opentelemetry.io/otel/sdk/export/trace"
-	"go.opentelemetry.io/otel/sdk/metric/controller/push"
-	"go.opentelemetry.io/otel/sdk/metric/processor/basic"
+	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
+	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/trace"
@@ -50,24 +52,26 @@ func NewExporter(options ...Option) (*Exporter, error) {
 // NewExportPipeline creates a complete export pipeline with the default
 // selectors, processors, and trace registration. It is the responsibility
 // of the caller to stop the returned push Controller.
-func NewExportPipeline(exportOpts []Option, pushOpts []push.Option) (trace.TracerProvider, *push.Controller, error) {
+func NewExportPipeline(exportOpts []Option, pushOpts []controller.Option) (trace.TracerProvider, *controller.Controller, error) {
 	exporter, err := NewExporter(exportOpts...)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	tp := sdktrace.NewTracerProvider(sdktrace.WithBatcher(exporter))
-	pusher := push.New(
-		basic.New(
+	pusher := controller.New(
+		processor.New(
 			simple.NewWithInexpensiveDistribution(),
 			exporter,
 		),
-		exporter,
-		pushOpts...,
+		append(
+			pushOpts,
+			controller.WithPusher(exporter),
+		)...,
 	)
-	pusher.Start()
+	err = pusher.Start(context.Background())
 
-	return tp, pusher, nil
+	return tp, pusher, err
 }
 
 // InstallNewPipeline creates a complete export pipelines with defaults and
@@ -82,7 +86,7 @@ func NewExportPipeline(exportOpts []Option, pushOpts []push.Option) (trace.Trace
 // 	}
 // 	defer pipeline.Stop()
 // 	... Done
-func InstallNewPipeline(exportOpts []Option, pushOpts []push.Option) (*push.Controller, error) {
+func InstallNewPipeline(exportOpts []Option, pushOpts []controller.Option) (*controller.Controller, error) {
 	tracerProvider, controller, err := NewExportPipeline(exportOpts, pushOpts)
 	if err != nil {
 		return controller, err